requirements.txt
requests>=2.25.1
python-dateutil>=2.8.2
pillow>=8.0.0